requirements.txt
# Core dependencies
openai>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
langchain>=0.0.308
langchain-openai>=0.0.2

# Browser automation 
browser-use>=0.2.4
playwright>=1.39.0

# Web scraping and parsing
lxml>=4.9.3
html2text>=2020.1.16

# Data processing
pydantic>=2.4.2
python-dateutil>=2.8.2

# Additional utilities
pillow>=10.0.0
tqdm>=4.66.1 